enhancing CRUD

+++ src/main/java/com/gabsthecreator/fullstackspringbootreact/student/StudentService.java
@@ -1,6 +1,8 @@
 package com.gabsthecreator.fullstackspringbootreact.student;
 
 
+import com.gabsthecreator.fullstackspringbootreact.student.exception.BadRequestException;
+import com.gabsthecreator.fullstackspringbootreact.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,16 +14,28 @@ public class StudentService {
 
     private final StudentRepository studentRepository;
 
-    public List<Student> getAllStudents(){
+    public List<Student> getAllStudents() {
         return studentRepository.findAll();
     }
 
     public void addStudent(Student student) {
-        // check if email is already taken
+        Boolean existsEmail = studentRepository.selectExistsEmail(student.getEmail());
+        if (existsEmail) {
+            throw new BadRequestException(
+                    "Sorry, the E-Mail " + student.getEmail() + " is already taken. Please use another one."
+            );
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
+        if (!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Sorry, a Student with ID " + studentId + " does not exist."
+            );
+        }
+
+
         studentRepository.deleteById(studentId);
     }
 }
